requirements.txt
accelerate==1.10.1
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
croniter==6.0.0
cryptography==45.0.6
deprecation==2.1.0
distro==1.9.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.7.0
h11==0.16.0
h2==4.2.0
hf-xet==1.1.9
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
lxml==6.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==1.99.1
packaging==25.0
plum-dispatch==1.7.4
postgrest==1.1.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
PyMuPDF==1.26.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
realtime==2.7.0
redis==6.4.0
regex==2025.7.34
requests==2.32.5
rq==2.6.0
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
spire-doc==13.8.0
starlette==0.47.2
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
sympy==1.14.0
tiktoken==0.11.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
